refactor show-client

+++ WebContent/WEB-INF/view/show-client.jsp
@@ -17,46 +17,42 @@
 		
 		<div id="container">
 			<div id="content">
+				
 				<table>
-						<tr>
-							<th>First Name</th>
-							<th>Last Name</th>
-							<th>Phone</th>
-							<th>Email</th>
-							<th>Formula</th>
-							<th>Action</th>
-						</tr>
-						<%-- <c:forEach var="tempClient" items="${client}"> --%>
-						
-						<c:url var="updateLink" value="/client/showFormUpdate">
-							<c:param name="clientId" value="${client.id}"/>
-						</c:url>
-						
-						<c:url var="deleteLink" value="/client/delete">
-							<c:param name="clientId" value="${client.id}"/>
-						</c:url>
-							<tr>
-								<td> ${client.firstName} </td>
-								<td> ${client.lastName} </td>
-								<td> ${client.phone} </td>
-								<td> ${client.email} </td>
-								<td> ${client.formula} </td>
-								<td> 
-									<a href="${updateLink}">Update</a>
-									|
-									<a href="${deleteLink}"
-									onclick="if (!(confirm('Are you sure you want to delete this client?'))) return false">Delete</a>
-								</td>
-							</tr>
-						<%-- </c:forEach> --%>	
-					</table>
+					<tr>
+						<th>First Name</th>
+						<th>Last Name</th>
+						<th>Phone</th>
+						<th>Email</th>
+						<th>Formula</th>
+						<th>Action</th>
+					</tr>
+					
+					<c:url var="updateLink" value="/client/showFormUpdate">
+						<c:param name="clientId" value="${client.id}"/>
+					</c:url>
+					<c:url var="deleteLink" value="/client/delete">
+						<c:param name="clientId" value="${client.id}"/>
+					</c:url>
+					
+					<tr>
+						<td> ${client.firstName} </td>
+						<td> ${client.lastName} </td>
+						<td> ${client.phone} </td>
+						<td> ${client.email} </td>
+						<td> ${client.formula} </td>
+						<td> 
+							<a href="${updateLink}">Update</a>
+							|
+							<a href="${deleteLink}" onclick="if (!(confirm('Are you sure you want to delete this client?'))) return false">Delete</a>
+						</td>
+					</tr>
+				</table>
 			</div>
-					<p>
-						<a href="${pageContext.request.contextPath}/client/list">Back To List</a>
-					</p>
+			
+			<p> <a href="${pageContext.request.contextPath}/client/list">Back To List</a> </p>
+		
 		</div>
-	
-	
 	</div>
 	
 </body>
